requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
